+++ src/main/java/janggi/domain/piece/generator/ChoPieceGenerator.java
@@ -1,9 +1,9 @@
-package janggi.domain.piece.gererator;
+package janggi.domain.piece.generator;
 
 import janggi.domain.piece.Piece;
 import java.util.List;
 
-public interface HanPieceGenerator {
+public interface ChoPieceGenerator {
 
     List<Piece> generate(KnightElephantSetting knightElephantSetting);
 }
