+++ src/test/java/janggi/domain/board/JanggiBoardTest.java
@@ -7,11 +7,11 @@ import janggi.domain.piece.Piece;
 import janggi.domain.piece.PieceType;
 import janggi.domain.piece.Position;
 import janggi.domain.piece.Side;
-import janggi.domain.piece.gererator.ChoPieceGenerator;
-import janggi.domain.piece.gererator.DefaultChoPieceGenerator;
-import janggi.domain.piece.gererator.DefaultHanPieceGenerator;
-import janggi.domain.piece.gererator.HanPieceGenerator;
-import janggi.domain.piece.gererator.KnightElephantSetting;
+import janggi.domain.piece.generator.ChoPieceGenerator;
+import janggi.domain.piece.generator.DefaultChoPieceGenerator;
+import janggi.domain.piece.generator.DefaultHanPieceGenerator;
+import janggi.domain.piece.generator.HanPieceGenerator;
+import janggi.domain.piece.generator.KnightElephantSetting;
 import janggi.domain.piece.movement.dynamic.RookMovementStrategy;
 import janggi.domain.piece.movement.fixed.ElephantMovementStrategy;
 import janggi.domain.piece.movement.fixed.KingMovementStrategy;
